requirements.txt
Django==3.2
Pillow==8.2.0
django-ckeditor==6.1.0
django-environ==0.4.5